+++ crates/test-plugin/src/lib.rs
@@ -1,7 +1,7 @@
 //! Plugin used for testing. We need a plugin with slightly different behavior
 //! to validate a plugin is actually used when it should be.
 
-use javy_plugin_api::{import_namespace, javy::Config};
+use javy_plugin_api::{import_namespace, Config};
 
 import_namespace!("test_plugin");